+++ src/vm/sequencers/native/op_create_map.go
@@ -41,7 +41,7 @@ func (op *OpCreateMap) Bind(vm core.IVM) error {
 func (op *OpCreateMap) Execute(decoder *core.Decoder) {
 	// Operands Offset 2 (16-bit)
 	numElements := decoder.Operand(0)
-	mElem := op.vm.StackPopMap(numElements)
+	mElem := op.vm.StackPopMap(uint(numElements))
 	op.vm.StackPush(op.vm.Factory().NewMap(op.vm.FrameId(), mElem))
 }
